Move colorstuff to static Colors class in colors.js
