+++ src/colors.js
@@ -1,0 +1,231 @@
+/** ***************************************************************************
+  * Colors class
+  *
+  * Summary.
+  *
+  */
+
+export default class Colors {
+  /** *****************************************************************************
+  * Colors::static properties()
+  *
+  * Summary.
+  * Declares the static class properties.
+  * Needs eslint parserOptions ecmaVersion: 2022
+  *
+  */
+  static {
+    Colors.colorCache = {};
+    Colors.element = undefined;
+  }
+
+  /** *****************************************************************************
+  * Colors::setElement()
+  *
+  * Summary.
+  * Sets the HTML element (the custom card) to work with getting colors
+  *
+  */
+
+  static setElement(argElement) {
+    Colors.element = argElement;
+  }
+
+  /** *****************************************************************************
+  * card::_calculateColor()
+  *
+  * Summary.
+  *
+  * #TODO:
+  * replace by TinyColor library? Is that possible/feasible??
+  *
+  */
+
+  static calculateColor(argState, argStops, argIsGradient) {
+    const sortedStops = Object.keys(argStops).map((n) => Number(n)).sort((a, b) => a - b);
+
+    let start; let end; let
+      val;
+    const l = sortedStops.length;
+
+    if (argState <= sortedStops[0]) {
+      return argStops[sortedStops[0]];
+    } else if (argState >= sortedStops[l - 1]) {
+      return argStops[sortedStops[l - 1]];
+    } else {
+      for (let i = 0; i < l - 1; i++) {
+        const s1 = sortedStops[i];
+        const s2 = sortedStops[i + 1];
+        if (argState >= s1 && argState < s2) {
+          [start, end] = [argStops[s1], argStops[s2]];
+          if (!argIsGradient) {
+            return start;
+          }
+          val = Colors.calculateValueBetween(s1, s2, argState);
+          break;
+        }
+      }
+    }
+    return Colors.getGradientValue(start, end, val);
+  }
+
+  /** *****************************************************************************
+  * card::_calculateColor2()
+  *
+  * Summary.
+  *
+  * #TODO:
+  * replace by TinyColor library? Is that possible/feasible??
+  *
+  */
+
+  static calculateColor2(argState, argStops, argPart, argProperty, argIsGradient) {
+    const sortedStops = Object.keys(argStops).map((n) => Number(n)).sort((a, b) => a - b);
+
+    let start; let end; let
+      val;
+    const l = sortedStops.length;
+
+    if (argState <= sortedStops[0]) {
+      return argStops[sortedStops[0]];
+    } else if (argState >= sortedStops[l - 1]) {
+      return argStops[sortedStops[l - 1]];
+    } else {
+      for (let i = 0; i < l - 1; i++) {
+        const s1 = sortedStops[i];
+        const s2 = sortedStops[i + 1];
+        if (argState >= s1 && argState < s2) {
+          // console.log('calculateColor2 ', argStops[s1], argStops[s2]);
+          [start, end] = [argStops[s1].styles[argPart][argProperty], argStops[s2].styles[argPart][argProperty]];
+          if (!argIsGradient) {
+            return start;
+          }
+          val = Colors.calculateValueBetween(s1, s2, argState);
+          break;
+        }
+      }
+    }
+    return Colors.getGradientValue(start, end, val);
+  }
+
+  /** *****************************************************************************
+  * card::_calculateValueBetween()
+  *
+  * Summary.
+  * Clips the argValue value between argStart and argEnd, and returns the between value ;-)
+  *
+  * Returns NaN if argValue is undefined
+  *
+  * NOTE: Rename to valueToPercentage ??
+  */
+
+  static calculateValueBetween(argStart, argEnd, argValue) {
+    return (Math.min(Math.max(argValue, argStart), argEnd) - argStart) / (argEnd - argStart);
+  }
+
+  /** *****************************************************************************
+  * card::_getColorVariable()
+  *
+  * Summary.
+  * Get value of CSS color variable, specified as var(--color-value)
+  * These variables are defined in the Lovelace element so it appears...
+  *
+  */
+
+  static getColorVariable(argColor) {
+    const newColor = argColor.substr(4, argColor.length - 5);
+
+    const returnColor = window.getComputedStyle(Colors.element).getPropertyValue(newColor);
+    return returnColor;
+  }
+
+  /** *****************************************************************************
+  * card::_getGradientValue()
+  *
+  * Summary.
+  * Get gradient value of color as a result of a color_stop.
+  * An RGBA value is calculated, so transparency is possible...
+  *
+  * The colors (colorA and colorB) can be specified as:
+  * - a css variable, var(--color-value)
+  * - a hex value, #fff or #ffffff
+  * - an rgb() or rgba() value
+  * - a hsl() or hsla() value
+  * - a named css color value, such as white.
+  *
+  */
+
+  static getGradientValue(argColorA, argColorB, argValue) {
+    const resultColorA = Colors.colorToRGBA(argColorA);
+    const resultColorB = Colors.colorToRGBA(argColorB);
+
+    // We have a rgba() color array from cache or canvas.
+    // Calculate color in between, and return #hex value as a result.
+    //
+
+    const v1 = 1 - argValue;
+    const v2 = argValue;
+    const rDec = Math.floor((resultColorA[0] * v1) + (resultColorB[0] * v2));
+    const gDec = Math.floor((resultColorA[1] * v1) + (resultColorB[1] * v2));
+    const bDec = Math.floor((resultColorA[2] * v1) + (resultColorB[2] * v2));
+    const aDec = Math.floor((resultColorA[3] * v1) + (resultColorB[3] * v2));
+
+    // And convert full RRGGBBAA value to #hex.
+    const rHex = Colors.padZero(rDec.toString(16));
+    const gHex = Colors.padZero(gDec.toString(16));
+    const bHex = Colors.padZero(bDec.toString(16));
+    const aHex = Colors.padZero(aDec.toString(16));
+
+    return `#${rHex}${gHex}${bHex}${aHex}`;
+  }
+
+  static padZero(argValue) {
+    if (argValue.length < 2) {
+      argValue = `0${argValue}`;
+    }
+    return argValue.substr(0, 2);
+  }
+
+  /** *****************************************************************************
+  * card::_colorToRGBA()
+  *
+  * Summary.
+  * Get RGBA color value of argColor.
+  *
+  * The argColor can be specified as:
+  * - a css variable, var(--color-value)
+  * - a hex value, #fff or #ffffff
+  * - an rgb() or rgba() value
+  * - a hsl() or hsla() value
+  * - a named css color value, such as white.
+  *
+  */
+
+  static colorToRGBA(argColor) {
+    // return color if found in colorCache...
+    const retColor = Colors.colorCache[argColor];
+    if (retColor) return retColor;
+
+    let theColor = argColor;
+    // Check for 'var' colors
+    const a0 = argColor.substr(0, 3);
+    if (a0.valueOf() === 'var') {
+      theColor = Colors.getColorVariable(argColor);
+    }
+
+    // Get color from canvas. This always returns an rgba() value...
+    const canvas = window.document.createElement('canvas');
+    // eslint-disable-next-line no-multi-assign
+    canvas.width = canvas.height = 1;
+    const ctx = canvas.getContext('2d');
+
+    ctx.clearRect(0, 0, 1, 1);
+    ctx.fillStyle = theColor;
+    ctx.fillRect(0, 0, 1, 1);
+    const outColor = [...ctx.getImageData(0, 0, 1, 1).data];
+
+    Colors.colorCache[argColor] = outColor;
+
+    return outColor;
+  }
+} // END OF CLASS
